Refactor to use WEM encryption service
Refactored the code to use the WEM encryption service instead of the previously used plenta encryption. This change affects multiple services that were using encryption directly from the system container. The Encryption service is now being injected into the classes that require it, improving code quality and testability. This change also removes dependency on plenta's contao-encryption in the composer.json.